Connecting the tasks to firebase
